requirements.txt
Flask
supabase
python-dotenv


